improve(webpack): Add filename to webpack config
Needed to add to overwrite the cases where the
imported config is setting filename to
be on a subdirectory like the case of
'/static/js/[name].js'. This would make the entry
file main.js be under that subdirectory instead of
at the top level like we assume. Overwritting it
always to 'main.js' will resolve this issue